# src/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cookie API",
    "description": "api description",
    "version": "1.0",
    "contact": {
      "name": "Vu Nguyen",
      "email": "aqaurius6666@gmail.com"
    }
  },
  "tags": [
    {
      "name": "CookieService"
    }
  ],
  "basePath": "/api",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/post-by-id": {
      "get": {
        "operationId": "CookieService_GetPostById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/swaggerGetPostByIdResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Post"
        ]
      }
    },
    "/posts": {
      "get": {
        "operationId": "CookieService_ListPosts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/swaggerListPostsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Post"
        ]
      },
      "post": {
        "operationId": "CookieService_CreatePosts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/swaggerCreatePostsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/swaggerCreatePostsRequest"
            }
          }
        ],
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{id}": {
      "put": {
        "operationId": "CookieService_UpdatePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/swaggerUpdatePostResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "isReceipe": {
                  "type": "boolean"
                },
                "tagIds": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "cookTime": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        ],
        "tags": [
          "Post"
        ]
      }
    },
    "/random-questions": {
      "get": {
        "operationId": "CookieService_RandomQuestions",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/swaggerRandomQuestionsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "number",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Post"
        ]
      }
    },
    "/suggestion-posts": {
      "get": {
        "operationId": "CookieService_SuggestionPosts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/swaggerSuggestionPostsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "tags",
            "description": "For example: tags[]=1\u0026tags[]=2\u0026tags[]=3",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Post"
        ]
      }
    },
    "/upload/presigned-upload": {
      "post": {
        "operationId": "CookieService_GetUploadUrl",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/swaggerGetUploadUrlResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/swaggerGetUploadUrlRequest"
            }
          }
        ],
        "tags": [
          "Upload"
        ]
      }
    }
  },
  "definitions": {
    "ListPostsResponsePostPagination": {
      "type": "object",
      "properties": {
        "pagination": {
          "$ref": "#/definitions/swaggerPagination"
        },
        "posts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/swaggerPost"
          }
        }
      }
    },
    "UploadUrlUploadHeader": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "swaggerCreatePostsRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "isReceipe": {
          "type": "boolean"
        },
        "tagIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "cookTime": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "swaggerCreatePostsResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/swaggerPost"
        }
      }
    },
    "swaggerGetPostByIdResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/swaggerPost"
        }
      }
    },
    "swaggerGetUploadUrlRequest": {
      "type": "object",
      "properties": {
        "fileName": {
          "type": "string"
        }
      }
    },
    "swaggerGetUploadUrlResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/swaggerUploadUrl"
        }
      }
    },
    "swaggerListPostsResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/ListPostsResponsePostPagination"
        }
      }
    },
    "swaggerPagination": {
      "type": "object",
      "properties": {
        "offset": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "swaggerPost": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "isReceipe": {
          "type": "boolean"
        },
        "author": {
          "$ref": "#/definitions/swaggerUser"
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/swaggerTag"
          }
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "upvote": {
          "type": "integer",
          "format": "int32"
        },
        "downvote": {
          "type": "integer",
          "format": "int32"
        },
        "cookTime": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "swaggerQuestion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "content": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/swaggerTag"
          }
        }
      }
    },
    "swaggerRandomQuestionsResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/swaggerQuestion"
          }
        }
      }
    },
    "swaggerSuggestionPostsResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/swaggerPost"
          }
        }
      }
    },
    "swaggerTag": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "swaggerUpdatePostResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/swaggerPost"
        }
      }
    },
    "swaggerUploadUrl": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        },
        "fields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UploadUrlUploadHeader"
          }
        }
      }
    },
    "swaggerUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}